LayoutTests/platform/glib/imported/w3c/web-platform-tests/xhr/setrequestheader-header-forbidden-expected.txt

CONSOLE MESSAGE: Refused to set unsafe header "Accept-Charset"
CONSOLE MESSAGE: Refused to set unsafe header "Accept-Encoding"
CONSOLE MESSAGE: Refused to set unsafe header "Connection"
CONSOLE MESSAGE: Refused to set unsafe header "Content-Length"
CONSOLE MESSAGE: Refused to set unsafe header "Cookie"
CONSOLE MESSAGE: Refused to set unsafe header "Cookie2"
CONSOLE MESSAGE: Refused to set unsafe header "Date"
CONSOLE MESSAGE: Refused to set unsafe header "DNT"
CONSOLE MESSAGE: Refused to set unsafe header "Expect"
CONSOLE MESSAGE: Refused to set unsafe header "Host"
CONSOLE MESSAGE: Refused to set unsafe header "Keep-Alive"
CONSOLE MESSAGE: Refused to set unsafe header "Referer"
CONSOLE MESSAGE: Refused to set unsafe header "TE"
CONSOLE MESSAGE: Refused to set unsafe header "Trailer"
CONSOLE MESSAGE: Refused to set unsafe header "Transfer-Encoding"
CONSOLE MESSAGE: Refused to set unsafe header "Upgrade"
CONSOLE MESSAGE: Refused to set unsafe header "Via"
CONSOLE MESSAGE: Refused to set unsafe header "Proxy-"
CONSOLE MESSAGE: Refused to set unsafe header "Proxy-LIES"
CONSOLE MESSAGE: Refused to set unsafe header "Proxy-Authorization"
CONSOLE MESSAGE: Refused to set unsafe header "Sec-"
CONSOLE MESSAGE: Refused to set unsafe header "Sec-X"

PASS XMLHttpRequest: setRequestHeader() - headers that are forbidden
